feat: Implement PostItem and PostList components with sample posts data
